Version API with namespacing
