+++ app/controllers/api/v1/authentication_controller.rb
@@ -1,0 +1,17 @@
+module Api
+  module V1
+    class AuthenticationController < ApplicationController
+      skip_before_action :authenticate_request
+
+      def authenticate
+        command = AuthenticateUser.call(params[:email], params[:password])
+
+        if command.success?
+          render json: { auth_token: command.result }
+        else
+          render json: command.errors, status: :unauthorized
+        end
+      end
+    end
+  end
+end
