--- app/controllers/authentication_controller.rb
@@ -1,13 +1,0 @@
-class AuthenticationController < ApplicationController
-  skip_before_action :authenticate_request
-
-  def authenticate
-    command = AuthenticateUser.call(params[:email], params[:password])
-
-    if command.success?
-      render json: { auth_token: command.result }
-    else
-      render json: command.errors, status: :unauthorized
-    end
-  end
-end
