+++ config/routes.rb
@@ -1,5 +1,9 @@
 Rails.application.routes.draw do
-  resources :users, except: [:index]
-  post 'auth', to: 'authentication#authenticate'
+  namespace :api do
+    namespace :v1 do
+      resources :users, except: [:index]
+      post 'auth', to: 'authentication#authenticate'
+    end
+  end
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
